benchmarks: split sweep orchestration out of cli.py
Move the ~340-line per-version provisioning block and both sweep bodies
into a new benchmarks/sweep.py (provision_venvs + run_sweep /
run_memory_sweep). cli.py's `sweep` and `memory sweep` commands become
thin shims that resolve their options (phase -> test file, smoke args)
and delegate. No behavior change — command set, flags, and help text are
identical; verified with a live one-version smoke sweep.

Per the plan's Item B but adapted in two ways:
- The shared-discovery helper goes into the existing snapshot.py as
  discover_snapshots() rather than a new snapshots.py — a sibling module
  one plural away from snapshot.py would be a nasty import footgun.
  _suggest_snapshots (typer-coupled presentation) stays in cli.py and
  calls it.
- run_memory_sweep moves too (not just run_sweep), so the provisioning
  generator stays private to sweep.py instead of being imported across a
  module boundary; all three memory subcommands are now thin.

cli.py: 1218 -> 882 lines (the remainder is command signatures + help).
