+++ benchmarks/cli.py
@@ -12,13 +12,9 @@ direct dispatch for registry introspection and memory snapshots.
 from __future__ import annotations
 
 import os
-import re
-import shutil
 import subprocess
 import sys
 import tempfile
-from collections.abc import Iterator
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Annotated, Literal
 
@@ -32,6 +28,8 @@ from benchmarks import (
 from benchmarks.memory import compare as memory_compare
 from benchmarks.memory import save as memory_save
 from benchmarks.plotting import FacetBy, Metric, PlotView, SortMode
+from benchmarks.snapshot import discover_snapshots
+from benchmarks.sweep import run_memory_sweep, run_sweep
 
 app = typer.Typer(
     help=(
@@ -50,23 +48,6 @@ app.add_typer(memory_app, name="memory")
 
 
 PhaseName = Literal["build", "matrices", "lp_write", "netcdf", "solver_handoff"]
-
-
-def _benchmarks_extra_pins() -> list[str]:
-    """
-    Return the pins from ``pyproject.toml``'s ``[benchmarks]`` extra.
-
-    Both ``sweep`` and ``memory sweep`` install these into each
-    per-version venv. Direct pins are kept in pyproject as the single
-    source of truth — bump them there and both sweeps pick up the
-    change. Transitive deps resolve fresh per venv; uv's deterministic
-    resolution gives identical results across versions within one sweep.
-    """
-    import tomllib
-
-    pyproject = Path(__file__).resolve().parents[1] / "pyproject.toml"
-    data = tomllib.loads(pyproject.read_text())
-    return list(data["project"]["optional-dependencies"]["benchmarks"])
 
 
 _PHASE_TEST_FILE: dict[PhaseName, str] = {
@@ -377,187 +358,6 @@ def notebook(
 # --- Sweep across linopy versions ------------------------------------------
 
 
-_PLAIN_VERSION_RE = re.compile(r"^\d+(\.\d+)*([a-z]+\d*)?$")
-
-
-def _linopy_install_spec(version: str) -> str:
-    """Turn ``0.4.0`` → ``linopy==0.4.0``, leave anything URL-y untouched."""
-    if _PLAIN_VERSION_RE.match(version):
-        return f"linopy=={version}"
-    return version
-
-
-def _venv_python(venv: Path) -> Path:
-    return (
-        venv / "Scripts" / "python.exe" if os.name == "nt" else venv / "bin" / "python"
-    )
-
-
-@dataclass(frozen=True)
-class _ProvisionedVenv:
-    """
-    One fresh per-version venv from :func:`_provision_venvs`.
-
-    On success, ``python``, ``env``, and ``import_dir`` are populated
-    and ``failed_at`` is ``None``. The caller MUST use ``import_dir``
-    as cwd for per-version subprocesses — see :func:`_provision_venvs`
-    for why. On failure, ``failed_at`` names the step that failed
-    (``"venv"``, ``"install"``, or ``"isolation"``); the caller skips
-    its per-version action and records the failure.
-    """
-
-    version: str
-    python: Path | None
-    env: dict[str, str] | None
-    import_dir: Path | None
-    failed_at: str | None
-
-
-def _provision_venvs(
-    versions: list[str], tmp_prefix: str, as_of: str | None = None
-) -> Iterator[_ProvisionedVenv]:
-    """
-    Yield one fresh per-version uv venv for each linopy version.
-
-    Used by both ``sweep`` and ``memory sweep`` so the venv plumbing
-    (uv venv → install ``[benchmarks]`` pins + the target linopy →
-    set up an isolated import root) lives in one place. The caller
-    supplies the tempdir prefix (so ``ps``/``lsof`` can distinguish
-    concurrent runs) and does whatever per-version action it needs.
-
-    **Isolation:** the repo root contains a ``linopy/`` package (the
-    one we're developing). Running the per-version pytest with the
-    repo root on ``sys.path`` — either via ``PYTHONPATH=repo`` or via
-    ``cwd=repo`` (Python prepends cwd as ``''``) — shadows the venv's
-    installed linopy with the dev tree. The whole sweep then measures
-    the dev linopy against itself instead of the requested version.
-    To avoid this, ``import_dir`` is a fresh tempdir per version that
-    holds a filtered *copy* of ``benchmarks/`` and nothing else — a
-    copy rather than a symlink so the sweep runs on Windows without
-    symlink privileges and so no per-version subprocess (nor its
-    ``__pycache__`` writes) ever touches the working tree. Running
-    subprocesses with ``cwd=import_dir`` and no ``PYTHONPATH`` makes
-    ``import benchmarks`` resolve to that copy while ``import linopy``
-    falls through to the venv's site-packages — i.e. the requested
-    version. The preflight below asserts that resolution actually held.
-
-    Each version's tempdir is cleaned up when the generator advances
-    (or exits). The caller can break the loop early — Python's
-    generator close protocol fires the ``with`` teardown.
-
-    **Cross-time reproducibility:** if ``as_of`` is a date string
-    (``YYYY-MM-DD`` or any ISO 8601 timestamp), passes
-    ``--exclude-newer`` to uv so the entire transitive resolution is
-    frozen to releases on or before that date. Pinning direct deps
-    alone (current default) keeps results reproducible *within* one
-    sweep call, but unpinned transitives can drift between sweep calls
-    days apart; ``as_of`` closes that gap.
-    """
-    if shutil.which("uv") is None:
-        typer.secho(
-            "uv not found on PATH — install via https://docs.astral.sh/uv/",
-            fg=typer.colors.RED,
-            err=True,
-        )
-        raise typer.Exit(code=2)
-
-    repo_root = Path.cwd()
-    for version in versions:
-        typer.secho(f"\n=== linopy {version} ===", fg=typer.colors.CYAN, bold=True)
-        with tempfile.TemporaryDirectory(prefix=tmp_prefix) as tmp:
-            venv = Path(tmp) / "venv"
-
-            r = subprocess.run(
-                ["uv", "venv", "--python", sys.executable, str(venv)],
-                check=False,
-            )
-            if r.returncode != 0:
-                typer.secho(
-                    f"venv creation failed: {version}",
-                    fg=typer.colors.RED,
-                    err=True,
-                )
-                yield _ProvisionedVenv(version, None, None, None, "venv")
-                continue
-
-            vpy = _venv_python(venv)
-            spec = _linopy_install_spec(version)
-
-            # Single install pass: pinned infra from pyproject + linopy.
-            # Direct pins in [benchmarks] are sufficient for sweep
-            # reproducibility — uv resolves the same input deterministically
-            # into each per-version venv.
-            install_args = [
-                "uv",
-                "pip",
-                "install",
-                "--python",
-                str(vpy),
-                *(["--exclude-newer", as_of] if as_of else []),
-                *_benchmarks_extra_pins(),
-                spec,
-            ]
-            r = subprocess.run(install_args, check=False)
-            if r.returncode != 0:
-                typer.secho(f"install failed: {version}", fg=typer.colors.RED, err=True)
-                yield _ProvisionedVenv(version, None, None, None, "install")
-                continue
-
-            # Build the isolated import root described in the docstring:
-            # a filtered copy of ``benchmarks/`` and nothing else. The
-            # heavy, sweep-irrelevant artifacts (the executed notebook,
-            # bytecode caches, macOS cruft) are skipped to keep the
-            # per-version copy cheap.
-            import_dir = Path(tmp) / "iso"
-            import_dir.mkdir()
-            shutil.copytree(
-                repo_root / "benchmarks",
-                import_dir / "benchmarks",
-                ignore=shutil.ignore_patterns("__pycache__", "*.ipynb", ".DS_Store"),
-            )
-
-            # No PYTHONPATH manipulation: the copied ``benchmarks`` under
-            # cwd=import_dir carries the harness without pulling the
-            # repo's ``linopy/`` into the import path. Bytecode the
-            # subprocess writes lands in this throwaway copy, never the
-            # working tree, so no PYTHONDONTWRITEBYTECODE is needed.
-            env = os.environ.copy()
-            env.pop("PYTHONPATH", None)
-
-            # Preflight: confirm the venv's linopy is what gets imported
-            # under cwd=import_dir. If a future change reintroduces the
-            # dev-linopy shadow bug, this fails loudly here rather than
-            # silently corrupting every snapshot in the sweep.
-            preflight = subprocess.run(
-                [
-                    str(vpy),
-                    "-c",
-                    (
-                        "import linopy; "
-                        f"assert {str(venv)!r} in linopy.__file__, "
-                        "f'isolation leak: linopy resolved to "
-                        "{linopy.__file__}, not the venv'"
-                    ),
-                ],
-                cwd=str(import_dir),
-                env=env,
-                capture_output=True,
-                text=True,
-                check=False,
-            )
-            if preflight.returncode != 0:
-                typer.secho(
-                    f"isolation preflight failed: {version}",
-                    fg=typer.colors.RED,
-                    err=True,
-                )
-                typer.echo(preflight.stderr.strip(), err=True)
-                yield _ProvisionedVenv(version, None, None, None, "isolation")
-                continue
-
-            yield _ProvisionedVenv(version, vpy, env, import_dir, None)
-
-
 @app.command(
     context_settings={"allow_extra_args": True, "ignore_unknown_options": True},
 )
@@ -663,119 +463,30 @@ def sweep(
     Wall-clock: roughly 1-2 minutes per version (venv + install +
     benchmarks). uv's wheel cache makes repeated runs much faster.
     """
-    if quick and long:
-        typer.secho(
-            "--quick and --long are mutually exclusive",
-            fg=typer.colors.RED,
-            err=True,
-        )
-        raise typer.Exit(code=2)
-
-    if smoke and (long or rounds is not None):
-        typer.secho(
-            "--smoke can't be combined with --long or --rounds "
-            "(no timings are recorded in smoke mode).",
-            fg=typer.colors.RED,
-            err=True,
-        )
-        raise typer.Exit(code=2)
-
-    if not smoke:
-        output_dir.mkdir(parents=True, exist_ok=True)
-
-    failed: list[str] = []
-    for prov in _provision_venvs(versions, "linopy-bench-", as_of=as_of):
-        if prov.failed_at:
-            failed.append(prov.version)
-            continue
-
-        if smoke:
-            # Smoke mode: reuse the same pytest args as the top-level
-            # ``smoke`` command. No JSON snapshot, return code is the
-            # signal.
-            pytest_cmd = [str(prov.python), "-m", "pytest", *_SMOKE_PYTEST_ARGS]
-            k_parts = [p for p in (model, filter_expr) if p]
-            if k_parts:
-                pytest_cmd.extend(["-k", " and ".join(k_parts)])
-            pytest_cmd.extend(ctx.args)
-
-            typer.secho(f"$ {' '.join(pytest_cmd)}", fg=typer.colors.BRIGHT_BLACK)
-            r = subprocess.run(
-                pytest_cmd, env=prov.env, cwd=str(prov.import_dir), check=False
-            )
-            if r.returncode != 0:
-                typer.secho(
-                    f"smoke failed: {prov.version}", fg=typer.colors.RED, err=True
-                )
-                failed.append(prov.version)
-            else:
-                typer.secho(f"smoke ok: {prov.version}", fg=typer.colors.GREEN)
-            continue
-
-        snapshot = (output_dir / f"linopy-{prov.version}.json").resolve()
-        test_target = _PHASE_TEST_FILE[phase] if phase is not None else "benchmarks/"
-        pytest_cmd = [
-            str(prov.python),
-            "-m",
-            "pytest",
-            test_target,
-            "--benchmark-only",
-            "--benchmark-json",
-            str(snapshot),
-        ]
-        if quick:
-            pytest_cmd.append("--quick")
-        elif long:
-            pytest_cmd.append("--long")
-        if rounds is not None:
-            pytest_cmd.extend(
-                [f"--benchmark-min-rounds={rounds}", "--benchmark-max-time=0"]
-            )
-
-        k_parts = [p for p in (model, filter_expr) if p]
-        if k_parts:
-            pytest_cmd.extend(["-k", " and ".join(k_parts)])
-
-        pytest_cmd.extend(ctx.args)
-
-        typer.secho(f"$ {' '.join(pytest_cmd)}", fg=typer.colors.BRIGHT_BLACK)
-        subprocess.run(pytest_cmd, env=prov.env, cwd=str(prov.import_dir), check=False)
-
-        if snapshot.exists():
-            typer.secho(f"saved {snapshot}", fg=typer.colors.GREEN)
-        else:
-            typer.secho(
-                f"no snapshot produced for {prov.version}",
-                fg=typer.colors.RED,
-                err=True,
-            )
-            failed.append(prov.version)
-
-    if failed:
-        typer.secho(f"\nFailed versions: {failed}", fg=typer.colors.RED, err=True)
-        raise typer.Exit(code=1)
+    test_target = _PHASE_TEST_FILE[phase] if phase is not None else "benchmarks/"
+    run_sweep(
+        versions,
+        output_dir=output_dir,
+        test_target=test_target,
+        smoke_args=_SMOKE_PYTEST_ARGS,
+        long=long,
+        quick=quick,
+        rounds=rounds,
+        model=model,
+        filter_expr=filter_expr,
+        smoke=smoke,
+        as_of=as_of,
+        extra_args=ctx.args,
+    )
 
 
 # --- Compare timing snapshots ---------------------------------------------
 
 
-def _discover_snapshots() -> list[Path]:
-    """
-    Return JSON snapshot files under the canonical .benchmarks/ tree.
-
-    Paths are relative to cwd so they're easier to copy-paste back into
-    the CLI than the absolute form would be.
-    """
-    root = Path(".benchmarks")
-    if not root.exists():
-        return []
-    return sorted(root.rglob("*.json"))
-
-
 def _suggest_snapshots(reason: str) -> None:
     """Print an error + a hint listing whatever snapshots we can find."""
     typer.secho(reason, fg=typer.colors.RED, err=True)
-    found = _discover_snapshots()
+    found = discover_snapshots()
     if found:
         typer.echo("\nAvailable snapshots under .benchmarks/:", err=True)
         for p in found:
@@ -1142,74 +853,14 @@ def memory_sweep_cmd(
     ``--repeats 1`` (default) is usually plenty. Use ``--repeats 3``
     if you need <5%% regression detection.
     """
-    from benchmarks.memory import MEMORY_PHASES
-
-    if phase:
-        unknown = [p for p in phase if p not in MEMORY_PHASES]
-        if unknown:
-            typer.secho(
-                f"unknown phase(s): {unknown}; valid options: {list(MEMORY_PHASES)}",
-                fg=typer.colors.RED,
-                err=True,
-            )
-            raise typer.Exit(code=2)
-
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    failed: list[str] = []
-    for prov in _provision_venvs(versions, "linopy-mem-", as_of=as_of):
-        if prov.failed_at:
-            failed.append(prov.version)
-            continue
-        # ``failed_at is None`` guarantees these are populated (see
-        # ``_ProvisionedVenv``); narrow for the type checker.
-        assert prov.python is not None and prov.import_dir is not None
-
-        # ``memory save`` writes to ``.benchmarks/memory/<label>.json``
-        # under cwd; we run it with cwd pinned to repo root, then move
-        # the file if the user asked for a custom output dir.
-        label = f"linopy-{prov.version}"
-        mem_cmd = [
-            str(prov.python),
-            "-m",
-            "benchmarks",
-            "memory",
-            "save",
-            label,
-        ]
-        if quick:
-            mem_cmd.append("--quick")
-        for ph in phase or []:
-            mem_cmd.extend(["--phase", ph])
-        if repeats > 1:
-            mem_cmd.extend(["--repeats", str(repeats)])
-
-        typer.secho(f"$ {' '.join(mem_cmd)}", fg=typer.colors.BRIGHT_BLACK)
-        subprocess.run(mem_cmd, env=prov.env, cwd=str(prov.import_dir), check=False)
-
-        # ``memory save`` writes to ``.benchmarks/memory/<label>.json``
-        # relative to its cwd — here, the isolated import_dir. Move it
-        # under the user's chosen output_dir (resolves under repo_root
-        # by default).
-        default_path = prov.import_dir / ".benchmarks" / "memory" / f"{label}.json"
-        target = output_dir / f"{label}.json"
-        if default_path.exists() and default_path.resolve() != target.resolve():
-            target.parent.mkdir(parents=True, exist_ok=True)
-            default_path.replace(target)
-
-        if target.exists():
-            typer.secho(f"saved {target}", fg=typer.colors.GREEN)
-        else:
-            typer.secho(
-                f"no snapshot produced for {prov.version}",
-                fg=typer.colors.RED,
-                err=True,
-            )
-            failed.append(prov.version)
-
-    if failed:
-        typer.secho(f"\nFailed versions: {failed}", fg=typer.colors.RED, err=True)
-        raise typer.Exit(code=1)
+    run_memory_sweep(
+        versions,
+        output_dir=output_dir,
+        quick=quick,
+        phases=phase,
+        repeats=repeats,
+        as_of=as_of,
+    )
 
 
 @memory_app.command("compare")
